% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2n.r
\name{chi2n}
\alias{chi2n}
\title{Balance test for sample size between treatments}
\usage{
chi2n(d, data)
}
\arguments{
\item{d}{a formula object whose right hand side is treatment variable (\code{~ d})}

\item{data}{an object whose class is \code{data.frame}}
}
\value{
this function returns a matrix including
sample size in each treatment and p-value of chi-squared test.
This matrix has \code{RCT} and \code{nbalance} classes.
}
\description{
This function implements chi-squared test for
the null hypothesis that sample size across treatments is equal.
}
