% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.r
\name{ctab}
\alias{ctab}
\title{A fucntion of getting coefficient table}
\usage{
ctab(x, ...)
}
\arguments{
\item{x}{a list or object including regression analysis}

\item{...}{other augments which pass on each augment.}
}
\description{
A fucntion of getting coefficient table
}
