% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbalance.r
\name{fbalance}
\alias{fbalance}
\title{Balance test of Covariates across treatments}
\usage{
fbalance(x, d, data)
}
\arguments{
\item{x}{a formula object whose right hand side is covariates (\code{~ x1 + x2}).}

\item{d}{a formula object whose right hand side is treatment variable (\code{~ d})}

\item{data}{an object whose class is \code{data.frame}}
}
\value{
A matrix including mean of covariate in each treatment and
p-value of F-statistics. This matrix has \code{RCT} and \code{xbalance} classes.
}
\description{
This function implements F-test for the null hypothesis that
observed variables are balanced across treatments.
First, this function runs a linear regression of covariate on treatments,
using \code{lm()}.
After that, we obtaine F-statistics and its p-value, using \code{anova()}.
}
