% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtemp.r
\name{ggtemp}
\alias{ggtemp}
\title{My ggplot2 template}
\usage{
ggtemp(flip = FALSE)
}
\arguments{
\item{flip}{logical value to whether \code{ggplot2::coord_flip()} is implemented.}
}
\value{
return an object whose class is \code{gg} and \code{theme}.
}
\description{
My ggplot2 template
}
