% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osterbias.r
\name{osterbias}
\alias{osterbias}
\title{Calculate Unobserved Selection Bias}
\usage{
osterbias(x, r_max = NULL, r_multiply = 1.3)
}
\arguments{
\item{x}{an object whose class is \code{RCT}.}

\item{r_max}{a numerical value of rmax. Default is null.}

\item{r_multiply}{a numerical value of multiple of r1. Default is \code{1.3}.}
}
\value{
a numeric vector whose class is \code{RCT} and \code{OsterBias}.
}
\description{
We calculate an effect of selection on unobservables
on estiamted coefficient of treatment. This method is presented by
Oster (2019, Journal of Business & Economic Statistics).
Let b0 and r0 be the coefficient of treatment and R-squared
resulting from the regression of outcome on treatment.
Let b1 and r1 be the coefficient of treatment and R-squared
resulting from the regression of outcome on
treatment and observed covariates.
Let rmax be the R-squared resulting from the hypothetical regression
of outcome on treatment, observed covairates, and unobserved covariates.
The value rmax depends on researchers.
Oster' suggestion is \code{rmax = r1 * 1.3}
(See a paper for detailed discussion).
Define b^ = b1 - (b0 - b1)*(rmax - r1)/(r1 - r0).
Under the following assumpetions,
the value b^ converges to a true b in probability one.
Assumption 1 is that
covariates are orthogonal to unobservables.
Assumption 2 is that
the unobservable and observables are equally related to the treatment.
Assumption 3 is that
the coefficient of covariates on treatment is same as
the coefficient of covairates on outcome.
These assumptions are not precise. See a paper for detailed discussion.
}
