% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.r
\name{reg}
\alias{reg}
\title{Unified regression analysis}
\usage{
reg(method, ...)
}
\arguments{
\item{method}{a srting of method of the generic function \code{reg}}

\item{...}{augments which pass on each method.}
}
\description{
This function implements various regression analysis,
especially econometrics.
This function has some advantages.
(1) Unifying augments.
By just changing method augment, you can use other regression analyses.
(2) Specifying each components of formula.
You can easily incorporate this function in a loop system.
}
\details{
There are four methods.
\subsection{\code{fe} method}{

Using the package lfe or fixest, you implement the fixed effect, and
the panel IV model. In addition to original class,
a returned object also has a class \code{fe_felm} (lfe pkg)
or \code{fe_fixest} (fixest pkg). Thus, you can use \code{summary()}.
You can pass following augments on this method:
\itemize{
\item \code{y}: a formula whose rhs is outcome (\code{y ~ .})
\item \code{x}: a formula whose lhs is covariates (\code{~ x1 + x2})
\item \code{z}: a formula whose rhs is endogenous variable
and lhs is exogenous variable (\code{x3 ~ z})
\item \code{fixef}: a formula whose lhs is fixed effects (\code{~ fix1 + fix2})
\item \code{cluster}: a formula whose lhs is cluster variable (\code{~ clust1})
\item \code{data}: a object whose class is \code{data.frame}.
\item \code{pkg}: a string of package. Default is \code{"fixest"}
\item \code{se}: method of caclulation of se (only \code{fixest} pkg).
the augment \code{se} passes a character scalar:
"standard", "hetero", "cluster", "twoway", "threeway" or "fourway."
By default if there are clusters in the estimation: se = "cluster";
otherwise se = "standard".
}
}

\subsection{\code{lpm} method}{

Using the \code{lm()} in the stats package,
you implement the linear proability model.
To claculate robust standard errors,
this method uses \code{coeftest} in the lmtest pkg,
and \code{vcovHC} in the sandwich pkg.
If treatment variable is specified,
you can estimate this model without covariates and with covariates.
A returned object is a list which includes \code{lm} result called \code{fit},
robust se test called \code{test},
and treatment variable formula called \code{treat} (if specified).
The returned object has \code{lpm} class in addition to original classes.
If treatment variable is specified, we add \code{RCT} class.
You can pass following augments on this method.
\itemize{
\item \code{y}: a formula whose rhs is outcome (\code{y ~ .})
\item \code{x}: a formula whose lhs is covariates (\code{~ x1 + x2})
\item \code{d}: a formula whose lhs is treatment variable (\code{~ d}).
\item \code{data}: an object whose class is \code{data.frame}
}
}

\subsection{\code{ols} method}{

Using the \code{lm} in the stats package,
you implement the linear regression model.
If treatment variable is specified,
you can estimate this model without covariates and with covariates.
A returned object is a list which includes \code{lm} result called \code{fit},
and treatment variable formula called \code{treat} (if specified).
The returned object has \code{ols} class in addition to original classes.
If treatment variable is specified, we add \code{RCT} class.
You can pass following augments on this method.
\itemize{
\item \code{y}: a formula whose rhs is outcome (\code{y ~ .})
\item \code{x}: a formula whose lhs is covariates (\code{~ x1 + x2})
\item \code{d}: a formula whose lhs is treatment variable (\code{~ d}).
\item \code{data}: an object whose class is \code{data.frame}
}
}

\subsection{\code{binomial} method}{

Using the \code{glm} in the stats package,
you implement the logit or probit model.
A returned object has \code{binomial} class in addition to original classes.
You can pass following augments on this method.
\itemize{
\item \code{y}: a formula whose rhs is outcome (\code{y ~ .})
\item \code{x}: a formula whose lhs is covariates (\code{~ x1 + x2})
\item \code{link}: character string: \code{"logit"} or \code{"probit"}.
\item \code{data}: an object whose class is \code{data.frame}
}
}
}
