% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.r
\name{regtab}
\alias{regtab}
\title{Make regression table}
\usage{
regtab(
  obj,
  keep_coef = NULL,
  rm_coef = NULL,
  label_coef = NULL,
  add_line = NULL,
  ...
)
}
\arguments{
\item{obj}{a list or object including regression analysis}

\item{keep_coef}{a character vector of variables kept}

\item{rm_coef}{a character vector of variables removed from table}

\item{label_coef}{a list including \verb{ols variable name = new variable name}}

\item{add_line}{a list including additional contents of each column.}

\item{...}{other augments which pass on each augment. See details.}
}
\value{
a data frame which includes all character strings
}
\description{
This function makes a character data frame of regression table.
An advantage point of this function is
that you can pass on output-format function such as \code{kable}.
}
\details{
the \code{binomial} method can show average marginal effect
if \code{ame = TRUE}, using \code{margins} in the \code{margins} package
}
