% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statvars.r
\name{se}
\alias{se}
\title{Calculate Standard Error of Mean}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numerical vector}

\item{na.rm}{a logical value whether we remove N/A value}
}
\value{
a numeric which is a statndard error of mean
}
\description{
Calculate Standard Error of Mean
}
