% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.r
\name{stab}
\alias{stab}
\title{A function of making regression stats table}
\usage{
stab(x)
}
\arguments{
\item{x}{a list or object including regression analysis}
}
\description{
A function of making regression stats table
}
