% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_rct.r
\name{ttest_rct}
\alias{ttest_rct}
\title{t-test across multiple treatment status}
\usage{
ttest_rct(y, d, base, data)
}
\arguments{
\item{y}{a formula object whose left hand side is outcome (\code{y ~ .}).}

\item{d}{a formula object whose right hand side is treatment variable (\code{~ d})}

\item{base}{a string or numeric value of the baseline group.}

\item{data}{an object whose class is \code{data.frame}}
}
\value{
A matrix including mean and its standard error of outcome
in each treatment and p-value of t-test.
This object has \code{RCT} and \code{ttest} class.
}
\description{
This function picks baseline group
from multiple treatment groups,
and implements t-test between the baseline group and
each treatment group.
}
